chore: rename and separate to tools
