+++ src/line/index.js
@@ -2,8 +2,7 @@
 // because I must refactor for draggable
 import Line from './Line';
 import LineTo from './LineTo';
-import Block from './Block';
 import SteppedLineTo from './SteppedLineTo';
 
 export default LineTo;
-export { SteppedLineTo, Line, Block };
+export { SteppedLineTo, Line };
